// src/app/api-services-v2/model/addressRequestDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddressRequestDTO { 
    postal_code?: string;
    building_name?: string;
    building_number?: string;
    comment?: string;
    country?: string;
    county?: string;
    street?: string;
    town: string;
}