// src/app/api-services-v2/model/customerIdResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerIdResponseDTO { 
    /**
     * ID des zugehörigen Lieferanten
     */
    supplier_id?: number;
    /**
     * Kundenkennung (extern oder intern)
     */
    customer_id?: string;
    /**
     * Optionaler Kommentar zur Kundenkennung
     */
    comment?: string;
}