// src/app/api-services-v2/model/supplierRequestDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressRequestDTO } from './addressRequestDTO';


export interface SupplierRequestDTO { 
    name: string;
    flag_preferred: boolean;
    vat_id?: string;
    email?: string;
    fax?: string;
    phone?: string;
    comment?: string;
    website?: string;
    address: AddressRequestDTO;
}